route bug

+++ routes/web.php
@@ -57,10 +57,10 @@ Route::group(['middleware' => ['auth']], function () {
     });
 
     Route::group(['middleware' => ['admin']], function () {
-        Route::get('/runningtext', [MarqueeController::class, 'runningtext'])->name('runningtext.index');
-        Route::post('/runningtext/store', [MarqueeController::class, 'store'])->name('runningtext.store');
-        Route::post('/runningtext/update{id}', [MarqueeController::class, 'update'])->name('runningtext.update');
-        Route::delete('/runningtext/delete{id}', [MarqueeController::class, 'destroy'])->name('runningtext.delete');
+        Route::get('admin/runningtext', [MarqueeController::class, 'adminrunningtext'])->name('admin.runningtext.index');
+        Route::post('admin/runningtext/store', [MarqueeController::class, 'adminstore'])->name('admin.runningtext.store');
+        Route::post('admin/runningtext/update{id}', [MarqueeController::class, 'adminadminupdate'])->name('admin.runningtext.update');
+        Route::delete('admin/runningtext/delete{id}', [MarqueeController::class, 'destroy'])->name('admin.runningtext.delete');
 
     });
 });
